--- client-runtime/src/main/java/com/microsoft/rest/v2/policy/AddHeaderPolicy.java
@@ -1,43 +1,0 @@
-/**
- * Copyright (c) Microsoft Corporation. All rights reserved.
- * Licensed under the MIT License. See License.txt in the project root for
- * license information.
- */
-
-package com.microsoft.rest.v2.policy;
-
-import com.microsoft.rest.v2.http.HttpRequest;
-import com.microsoft.rest.v2.http.HttpResponse;
-import rx.Single;
-
-/**
- * Type representing RequestPolicy that can intercept and modify headers.
- */
-public final class AddHeaderPolicy implements RequestPolicy {
-    private final RequestPolicy next;
-
-    /**
-     * Creates AddHeaderPolicy.
-     *
-     * @param next reference to the next RequestPolicy in the request-response pipeline.
-     */
-    public AddHeaderPolicy(RequestPolicy next) {
-        this.next = next;
-    }
-
-    /**
-     * Factory to create AddHeaderPolicy.
-     */
-    static class Factory implements RequestPolicy.Factory {
-        @Override
-        public RequestPolicy create(RequestPolicy next) {
-            return new AddHeaderPolicy(next);
-        }
-    }
-
-    @Override
-    public Single<HttpResponse> sendAsync(HttpRequest request) {
-        request.headers().add("x-my-header", "42");
-        return next.sendAsync(request);
-    }
-}
